Added Freemarker to WeBattalia. Added Spark package for FreemarkerEngine.
In src/main/resources added folders for external resourses.
The css, png, and svg folders are to be hosted externally.
The templates folder is to reside on the same level as the .jar file.

Edited Main to now display a Freemarker built webpage with a starting
hand using resources hosted at nickmacholl.com.

Changed some code in the game package for consistent naming.