Added CardType and modified Cost
Added CardType enumeration and modified Cost to include a CardType and a
priceList determined by the cardType. Removed name from Card and
replaced with type. 
Cost includes a method to pass in a hand and determine whether the
priceList is contained within that hand. This allows the user to
determine whether a particular unit can be purchased by a hand. The test
of this code is included in DeckTest.