Add City class and subclasses. Add subclasses to Road
This update adds the subclasses Straight, Curved, Tee, and Cross to the
Road class. These represent the types of roads in the game. The City
class inherits from the Road class and also has Straight, Curved, Tee,
and Cross as well as End to represent the five types of cities. The City
class has an int level and a Faction owner property. It has a special
Cost(int level) class for finding the different costs for building a
city as well as an upgradeCost() class that gives the cost of upgrading
a city of level 1, 2, or 3.