Created Market Class and CardRegistry Class. Turned Oracle to Singleton
This update has a number of updates. The Market Class initializes all
the decks used in the game. The CardRegistry adds ids to each card that
is created. It is a singleton so it can be called using
CardRegistry.getInstance(). The Oracle is also a singleton and is called
using Oracle.getInstance().